requirements.txt
pandas==1.4.4
plotly==5.10.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
streamlit==1.12.2
binance-connector==1.17.0
